Make status field immutable except via state transitions
Enable protected=True on the status fields of Order and OrderItem,
making the field immutable except via one of the defined state
transitions.

This means that these models can no longer use refresh_from_db(), as
explained in the Django-FSM documentation:

https://github.com/viewflow/django-fsm#usage

Therefore, fix up the tests accordingly so that rather than refreshing
an object, the reinitialize it with a database lookup using the
primary key.