Allow looking up the course ID from the SKU, via a redirect
The general expectation is that the system invoking the webhook
will have the product SKU configured to an Open edX course ID.

Sometimes, however, that is not an option, or at least is not
desired. For example, an Open edX platform may have several
consecutive course runs that from a commercial perspective are all one
and the same product. Thus, allow platform administrators to apply the
following hack:

* Open edX exposes the Django redirects app, which allows platform
  administrators to define arbitrary HTTP redirects.

* An admin now has the option to define a redirect that allows a
  course lookup by SKU, and that redirects to a full course URL. For
  example, they might use https://my.lms.domain/sku/xyz123 to redirect
  to https://my.lms.domain/courses/course-v1:org+course+run/.

* What the webhook receiver will now do, if it detects a SKU that does
  not start with "course-v1:", is do an HTTP HEAD request (with
  redirects enabled), against https://my.lms.domain/$prefix$sku (where
  $prefix is configurable), and extract the course ID from the
  location it is being redirected to.