Add the "mail_admins" logging handler to the production settings
Since receiving webhooks is potentially a business-critical process,
we want to give administrators the ability to set the ADMINS and
DEFAULT_FROM_EMAIL options and receive email messages any time we run
into an ERROR condition.

To that end, define a "mail_admins" handler and add it to the "django"
and "django.requests" loggers. Also add it to the "requests" logger,
to catch errors in interactions with the LMS REST API.

This handler does nothing if

* we're in DEBUG mode,
* ADMINS is unset or empty.