Support the WooCommerce dummy webhook invocation
When WooCommerce web hooks are first created or enabled, WooCommerce
sends a POST request whose content type is not application/json, but
instead application/x-www-form-urlencoded with a single form value:
"webhook_id=<num>".

This request does not contain a signature nor a valid source header,
so there is little we can do to verify that the webhook is properly
signed. But WooCommerce will show an error if the test POST to the
webhook returns anything but HTTP 200.

So, simply check for the content type and the presence of the
webhook_id form value, and return HTTP 200 to keep the WooCommerce
admin happy. If we receive something that is
application/x-www-form-urlencoded but does not contain webhook_id, log
a warning message instead.