Enable requiring payment for WooCommerce orders
WooCommerce, unlike Shopify, fires the order.created webhook *before*
the customer has completed the payment process. Processing the webhook
at that stage (and thus enrolling a learner in a course before they
have completed payment) is usually premature.

WooCommerce does not have a separate event that fires when payment on
an order has completed; it just fires a generic order.updated webhook
— which, howver, does contain all the information that order.created
contains. We thus apply the following heuristic:

If we require payment on an order (through the require_payment toggle)
and the webhook payload has date_paid_gmt set to null (i.e. the order
is unpaid), we store the webhook data, but do not process the
order. Only when we get a webhook that has date_paid_gmt set to
something that is parseable as a date, do we actually process the
order and create an enrollment.