Raise exception to prevent empty database config
env.db() returns an empty database config dictionary if
DJANGO_DATABASE_URL is set to an empty string. This will break when
we want to override default database configuration parameters for
production. Raise an exception and inform the user to either send in
a correct URL or remove the DJANGO_DATABASE_URL environment variable
so that the default sqlite configuration is used to populate the
config dict.

Also switch to django-environ when reading in DB_MIGRATION_* values to
ensure proper types and to use the same method of reading envvars in
all setting files.