Fix migrations broken in MySQL
Migrations introduced in #12 fail in MySQL due to the bug discussed
in https://code.djangoproject.com/ticket/31335.

In short, removing the UniqueConstraint fails in MySQL
as it cannot drop the index that is used in the "order" field
foreign key constraint.

The workaround is to remove the foreign key constraint first,
remove the old UniqueConstraint, add the new one, and then add
the foreign key constraint again using AlterField on "order".