Account for list of strings in WooCommerce line item metadata
In WooCommerce JSON, in the meta_data field of a list item in
line_items, also ignore the value if it's a list of strings and not
only when it's a single string.

In addition, log detected email addresses and ignored metadata at the
DEBUG level (which is normally suppressed).